Add files via upload
This data talks about the songs that are available on the Spotify platform. It starts by determining the playlist of the song. I tell you this about other data, including the type of song and the name. When I downloaded the song and the song has a movement first, and from these data I select the playlist (playlist_genre)